Requests and responses are **protocol-independent**
