+++ src/main/java/net/redstonelamp/Server.java
@@ -23,7 +23,7 @@ import net.redstonelamp.network.Protocol;
 import net.redstonelamp.network.pc.PCProtocol;
 import net.redstonelamp.network.pe.PEProtocol;
 import net.redstonelamp.plugin.PluginSystem;
-import net.redstonelamp.request.pe.v27.LoginRequest;
+import net.redstonelamp.request.LoginRequest;
 import net.redstonelamp.response.Response;
 import net.redstonelamp.ticker.RedstoneTicker;
 import net.redstonelamp.ui.Logger;
