+++ src/main/java/net/redstonelamp/request/PlayerMoveRequest.java
@@ -14,16 +14,26 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with RedstoneLamp.  If not, see <http://www.gnu.org/licenses/>.
  */
-package net.redstonelamp.response.pe.v27;
+package net.redstonelamp.request;
 
-import net.redstonelamp.response.Response;
+import net.redstonelamp.level.position.Position;
 
-public class PopupResponse extends Response{
+/**
+ * A Request from the player to move.
+ *
+ * @author RedstoneLamp Team
+ */
+public class PlayerMoveRequest extends Request{
+    public Position position;
+    public boolean onGround;
 
-    public String message;
-
-    public PopupResponse(String message){
-        this.message = message;
+    public PlayerMoveRequest(Position position, boolean onGround){
+        this.position = position;
+        this.onGround = onGround;
     }
 
+    @Override
+    public void execute(){
+        //TODO?
+    }
 }
