+++ src/main/java/net/redstonelamp/response/PlayerMoveResponse.java
@@ -14,24 +14,25 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with RedstoneLamp.  If not, see <http://www.gnu.org/licenses/>.
  */
-package net.redstonelamp.response.pe.v27;
+package net.redstonelamp.response;
 
 import net.redstonelamp.level.position.Position;
-import net.redstonelamp.response.Response;
 
 /**
- * A Response that teleports the player to the given position.
+ * A Response to a Player Move Request.
  *
  * @author RedstoneLamp Team
  */
-public class TeleportResponse extends Response{
+public class PlayerMoveResponse extends Response{
     public Position pos;
     public float bodyYaw;
     public boolean onGround;
+    public long entityID;
 
-    public TeleportResponse(Position pos, boolean onGround){
+    public PlayerMoveResponse(long eid, Position pos, boolean onGround){
         this.pos = pos;
         this.onGround = onGround;
         bodyYaw = pos.getYaw();
+        entityID = eid;
     }
 }
