+++ src/main/java/net/redstonelamp/response/StatusResponse.java
@@ -1,0 +1,5 @@
+package net.redstonelamp.response;
+
+public class StatusResponse{
+
+}
