--- src/main/java/net/redstonelamp/response/pc/StatusResponse.java
@@ -1,5 +1,0 @@
-package net.redstonelamp.response.pc;
-
-public class StatusResponse{
-
-}
